backend/requirements.txt
fastapi
uvicorn
pydantic
firebase-admin
google-cloud-firestore
numpy
pandas
python-dotenv
joblib
scikit-learn
python-multipart
# Add other dependencies here
